pyproject.toml
[build-system]
requires = ["setuptools >= 68.0", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "mreg"
requires-python = ">=3.10"
dependencies = [
    "Django>=5.0a1,<5.1",
    "djangorestframework==3.14.0",
    "django-auth-ldap>=5.1.0",
    "django-logging-json>=1.15",
    "django-netfields>=1.3.2",
    "django-filter>=24.3",
    "structlog>=24.4.0",
    "rich>=13.9.4",
    "gunicorn>=23.0.0",
    "idna>=3.10",
    "psycopg2-binary>=2.9.10",
    "pika>=1.3.2",
    "sentry-sdk>=2.8.0",
    "tzdata>=2024.2",
    # For OpenAPI schema generation.
    "uritemplate",
    "pyyaml",
    # For testing inside Docker image
    "unittest-parametrize"
]
dynamic = ["version"]

# https://docs.astral.sh/uv/concepts/dependencies/#dependency-groups
[dependency-groups]
dev = [
    "tox-uv",
    "coverage[toml]",
    "pytest",
    "pytest-django",
]
ci = [
    # Explictly include dev group for non-uv package managers
    # (dev group is automatically installed with uv)
    {include-group = "dev"},
    "tox-gh-actions", 
    "coveralls",
]

[tool.ruff]
line-length = 140
exclude = ["mreg/migrations/", "hostpolicy/migrations/", ".tox"]

[tool.ruff.lint]
# https://beta.ruff.rs/docs/rules/
select = ["E", "F"]

[tool.setuptools]
py-modules = ["mreg", "mregsite", "hostpolicy"]

[tool.setuptools_scm]
version_file = "mreg/_version.py"
